[fpmsyncd]: Add nexthop definition and decoding aligning with zebra
After receiving the nexthop group's full information from zebra,
we decode the message and construct a NextHopGroupFull to store
the attributes, which are generally aligned with zebra.

NextHopGroupFull is used for both multi-path and singleton NHG.
For multi-path NHG, it contains the depends and dependents id array,
and for singleton NHG, it contains the explicit attribute values.

NextHopGroupFull finished constructing will be passed to FIB BLOCK
to update the memory mapping table and APP_DB.

Signed-off-by: Yuqing Zhao <galadriel.zyq@alibaba-inc.com>